Revert "feat: eagerly connect outside of react lifecycle (#7334)"
This reverts commit 54880d201a3629988868dcd80c87a7794876cf44.
