+++ cypress/support/commands.ts
@@ -53,6 +53,7 @@ Cypress.Commands.overwrite(
 
             setInitialUserState(win, {
               ...initialState,
+              hideUniswapWalletBanner: true,
               ...CONNECTED_WALLET_USER_STATE,
               ...(options?.userState ?? {}),
             })
