+++ cypress/utils/user-state.ts
@@ -10,25 +10,23 @@ export const DISCONNECTED_WALLET_USER_STATE: Partial<UserState> = { selectedWall
  * Other persisted slices are not set, so they will be filled with their respective initial values
  * when the app runs.
  */
-export function setInitialUserState(win: Cypress.AUTWindow, state: UserState) {
-  // Selected wallet should also be reflected in localStorage, so that eager connections work.
-  if (state.selectedWallet) {
-    win.localStorage.setItem('selected_wallet', state.selectedWallet)
-  }
-
+export function setInitialUserState(win: Cypress.AUTWindow, initialUserState: any) {
   win.indexedDB.deleteDatabase('redux')
+
   const dbRequest = win.indexedDB.open('redux')
+
   dbRequest.onsuccess = function () {
     const db = dbRequest.result
     const transaction = db.transaction('keyvaluepairs', 'readwrite')
     const store = transaction.objectStore('keyvaluepairs')
     store.put(
       {
-        user: state,
+        user: initialUserState,
       },
       'persist:interface'
     )
   }
+
   dbRequest.onupgradeneeded = function () {
     const db = dbRequest.result
     db.createObjectStore('keyvaluepairs')
