+++ src/connection/types.ts
@@ -2,8 +2,6 @@ import { ChainId } from '@uniswap/sdk-core'
 import { Web3ReactHooks } from '@web3-react/core'
 import { Connector } from '@web3-react/types'
 
-export const selectedWalletKey = 'selected_wallet'
-
 export enum ConnectionType {
   UNISWAP_WALLET_V2 = 'UNISWAP_WALLET_V2',
   INJECTED = 'INJECTED',
@@ -11,14 +9,6 @@ export enum ConnectionType {
   WALLET_CONNECT_V2 = 'WALLET_CONNECT_V2',
   NETWORK = 'NETWORK',
   GNOSIS_SAFE = 'GNOSIS_SAFE',
-}
-
-export function toConnectionType(value = ''): ConnectionType | undefined {
-  if (Object.keys(ConnectionType).includes(value)) {
-    return value as ConnectionType
-  } else {
-    return undefined
-  }
 }
 
 export interface Connection {
